requirements.txt
atproto
python-dotenv
ollama
